update RepoPage to render bundles tab
